requirements.txt
aiogram>=3.0.0
python-dotenv
pandas
openpyxl
torch
transformers
Pillow
psycopg2-binary
